wip(rfj): implémentation des méthodes delete et update pour valider les tests. Gestion réponses REST

+++ src/main/java/ch/hearc/cafheg/business/allocations/AllocataireService.java
@@ -3,19 +3,26 @@ package ch.hearc.cafheg.business.allocations;
 import ch.hearc.cafheg.infrastructure.api.dto.AllocataireDTO;
 import ch.hearc.cafheg.infrastructure.api.dto.AllocataireDTOMapper;
 import ch.hearc.cafheg.infrastructure.persistance.AllocataireMapper;
+import ch.hearc.cafheg.infrastructure.persistance.VersementMapper;
 
 public class AllocataireService {
 
     private final AllocataireMapper allocataireMapper;
     private final AllocataireDTOMapper allocataireDTOMapper;
+    private final VersementMapper versementMapper;
 
-    public AllocataireService(AllocataireMapper allocataireMapper, AllocataireDTOMapper allocataireDTOMapper) {
+    public AllocataireService(AllocataireMapper allocataireMapper, AllocataireDTOMapper allocataireDTOMapper, VersementMapper versementMapper) {
         this.allocataireMapper = allocataireMapper;
         this.allocataireDTOMapper = allocataireDTOMapper;
+        this.versementMapper = versementMapper;
     }
 
     public String deleteById(int allocataireId) {
-        return allocataireMapper.deleteById(allocataireId);
+        if (versementMapper.countVersementsByAllocataireId(allocataireId) == 0) {
+            allocataireMapper.deleteById(allocataireId);
+            return "Allocataire supprimé";
+        }
+        return "Pas possible de supprimer";
     }
 
     public AllocataireDTO update(AllocataireDTO allocataireDTO) {
